refactor: Move player related things to the player module
Visibility changes to `pub`:
- some fields in `Player`
